Moved loginUser action to loginUserActions.js and added username and token to mapStateToProps
