+++ src/actions/loginUserActions.js
@@ -1,0 +1,39 @@
+// Action Types
+export const AUTH_USER = 'AUTH_USER';
+export const UNAUTH_USER = 'UNAUTH_USER';
+export const AUTH_ERROR = 'AUTH_ERROR';
+
+// Action Creators
+export const loginUser  = (credentials) => {
+  
+    return function(dispatch){
+        console.log("credentials: ", credentials);
+
+        const url = "https://kwitter-api.herokuapp.com/auth/login";
+
+        const postRequestOptions = {
+            method: "POST",
+            headers: {
+                "Content-Type": "application/json"
+            },
+            body: JSON.stringify({credentials}),
+        }
+
+        fetch(url, postRequestOptions)
+        .then(response => response.json())
+        .then(data => {
+            console.log("data: ", data);
+            dispatch({ type: AUTH_USER,
+                       username: credentials.username,
+                       token: data.token  });
+        }).catch(error => {
+            return error;
+        });
+    }
+  }
+
+  export const logoutUser = () => {  
+    return function (dispatch) {
+      dispatch({ type: UNAUTH_USER });
+    }
+  }
