+++ src/index.js
@@ -7,7 +7,8 @@ import './index.css';
 import App from './App';
 import registerServiceWorker from './registerServiceWorker';
 import thunk from 'redux-thunk';
-import reducer from './reducers';
+import reducer from './reducers'
+// import rootReducer from './reducers';
 
 const store = createStore(reducer, 
     compose(applyMiddleware(thunk),
